requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7
pyotp==2.9.0

# Redis (Caching & Sessions)
redis==5.0.1
hiredis==2.2.3

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Monitoring & Logging
structlog==23.2.0
python-json-logger==2.0.7
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0
python-dateutil==2.8.2

# POPIA Compliance
faker==20.1.0  # For generating test data (anonymized)

